iteration 2: move io logic to separate component and separate test;
